requirements_new.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
tkinter
pytest==7.4.3
